Move installHelpers
