+++ install.js
@@ -10,7 +10,7 @@ var readline = require('readline');
 var auth = require('./lib/auth');
 var database = require('./lib/database');
 var helpers = require('./lib/helpers');
-var installHelpers = require('./installHelpers');
+var installHelpers = require('./lib/installHelpers');
 var localAuth = require('./plugins/auth/local');
 var logger = require('./lib/logger');
 var origin = require('./lib/application');
