+++ lib/application.js
@@ -25,7 +25,7 @@ var EventEmitter = require('events').EventEmitter,
     Mailer = require('./mailer').Mailer,
     configuration = require('./configuration');
 
-var installHelpers = require('../installHelpers')
+var installHelpers = require('./installHelpers')
 
 var async = require('async');
 var chalk = require('chalk');
