+++ lib/bowermanager.js
@@ -8,7 +8,7 @@ var path = require('path');
 var rimraf = require('rimraf');
 var semver = require('semver');
 
-var installHelpers = require('../installHelpers');
+var installHelpers = require('./installHelpers');
 var Constants = require('./outputmanager').Constants;
 var database = require('./database');
 var logger = require('./logger');
