+++ lib/installHelpers.js
@@ -8,7 +8,7 @@ var request = require('request');
 var semver = require('semver');
 var Spinner = require('cli-spinner').Spinner;
 
-var configuration = require('./lib/configuration');
+var configuration = require('./configuration');
 
 var DEFAULT_USER_AGENT = 'Mozilla/5.0 (Windows NT 6.1; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/41.0.2272.118 Safari/537.36';
 var DEFAULT_SERVER_REPO = 'https://github.com/adaptlearning/adapt_authoring.git';
