+++ lib/tenantmanager.js
@@ -11,7 +11,7 @@ var util = require('util');
 
 var configuration = require('./configuration');
 var database = require('./database');
-var installHelpers = require('../installHelpers');
+var installHelpers = require('./installHelpers');
 var logger = require('./logger');
 
 // Constants
